+++ backend/src/proto/mod.rs
@@ -6,8 +6,8 @@ pub struct OrderRequest {
     pub user_id: String,
     #[prost(string, tag = "2")]
     pub symbol: String,
-    #[prost(uint32, tag = "3")]
-    pub quantity: u32,
+    #[prost(double, tag = "3")]
+    pub quantity: f64,
     #[prost(uint32, tag = "4")]
     pub side: u32,
     #[prost(string, tag = "5")]
@@ -20,28 +20,28 @@ pub struct LimitOrderRequest {
     pub user_id: String,
     #[prost(string, tag = "2")]
     pub symbol: String,
-    #[prost(uint32, tag = "3")]
-    pub quantity: u32,
+    #[prost(double, tag = "3")]
+    pub quantity: f64,
     #[prost(uint32, tag = "4")]
     pub side: u32,
-    #[prost(uint64, tag = "5")]
-    pub price: u64,
+    #[prost(double, tag = "5")]
+    pub price: f64,
     #[prost(string, tag = "6")]
     pub order_type: String,
-    #[prost(uint64, optional, tag = "7")]
-    pub tp: Option<u64>,
-    #[prost(uint64, optional, tag = "8")]
-    pub sl: Option<u64>,
+    #[prost(double, optional, tag = "7")]
+    pub tp: Option<f64>,
+    #[prost(double, optional, tag = "8")]
+    pub sl: Option<f64>,
 }
 
 #[derive(Clone, PartialEq, Message)]
 pub struct ModifyOrderRequest {
     #[prost(string, tag = "1")]
     pub user_id: String,
-    #[prost(uint64, tag = "2")]
-    pub tp: u64,
-    #[prost(uint64, tag = "3")]
-    pub sl: u64,
+    #[prost(double, tag = "2")]
+    pub tp: f64,
+    #[prost(double, tag = "3")]
+    pub sl: f64,
 }
 
 #[derive(Clone, PartialEq, Message)]
@@ -56,16 +56,16 @@ pub struct CloseOrderRequest {
 pub struct OrderResponse {
     #[prost(string, tag = "1")]
     pub message: String,
-    #[prost(uint32, tag = "2")]
-    pub quantity: u32,
+    #[prost(double, tag = "2")]
+    pub quantity: f64,
 }
 
 #[derive(Clone, PartialEq, Message)]
 pub struct CloseOrderResponse {
     #[prost(string, tag = "1")]
     pub message: String,
-    #[prost(uint64, tag = "2")]
-    pub close_price: u64,
+    #[prost(double, tag = "2")]
+    pub close_price: f64,
 }
 
 #[derive(Clone, PartialEq, Message)]
